Begin refactor project loading and add legacy (RPG2k/2k3) project loading.
Instead of handling everything in Core a dedicated project class was created which handles everything related to the project.
This is the initial step to achieve this, too much logic is still in core and main_window.

The well-known FileFinder was added but it is much less complex than the Player FileFinder because we can use Qt API here.