Refactor the Views one last time.
The GraphicsItem generator in the ActorModel was not flexible enough (no way to configure walking, facing, etc. easily).
I reused now the existing CharItem/FaceItem and migrated the code from the model into it. This is the best of both worlds.

Also replaced the GraphicsViews with a template as all the code is shared.